Disable tracking assignment location Fix declaration of variables in for loop
