+++ src/Analysis/Ast/Impl/Linting/UndefinedVariables/ExpressionWalker.cs
@@ -79,14 +79,20 @@ namespace Microsoft.Python.Analysis.Linting.UndefinedVariables {
             // undefined x in 
             //    y = x
             //    x = 1
-            if(m is ILocatedMember lm && lm.Location.DocumentUri == _analysis.Document.Uri) {
-                if (!(m is IPythonFunctionType || m is IPythonClassType)) {
-                    var span = lm.Location.Span;
-                    var nodeLoc = node.GetLocation(_analysis.Document);
-                    if (span.IsAfter(nodeLoc.Span)) {
-                        _analysis.ReportUndefinedVariable(node);
-                    }
+            if (m is ILocatedMember lm && lm.Location.DocumentUri == _analysis.Document.Uri) {
+                // Do not complain about functions and classes that appear later in the file
+                if (m is IPythonFunctionType || m is IPythonClassType) {
+                    return false;
                 }
+                // Since analyzer remembers the last assignment, it may look that the variable
+                // is not defined yet, but we also should look up, perhaps it was defined
+                // earlier and redefined later.
+                // https://github.com/Microsoft/python-language-server/issues/682
+                //var span = lm.Location.Span;
+                //var nodeLoc = node.GetLocation(_analysis.Document);
+                //if (span.IsAfter(nodeLoc.Span)) {
+                //    _analysis.ReportUndefinedVariable(node);
+                //}
             }
             return false;
         }
